requirements.txt
# Core ML/AI Libraries
faster-whisper==1.0.3
google-genai==1.45.0
TTS==0.22.0

# Audio Processing
sounddevice==0.4.6
numpy==1.24.3
pygame==2.5.2

# Web Interface
gradio==4.44.0

# Utilities
python-dotenv==1.0.0
